refactor: Reception을 view의 역할로 이동
