+++ src/main/kotlin/lottery/controller/LotteryStore.kt
@@ -1,13 +1,10 @@
 package lottery.controller
 
-import lottery.domain.Profit
-import lottery.domain.BonusBall
-import lottery.domain.LotteryFactory
-import lottery.domain.LotteryMatcher
-import lottery.domain.WinnerLottery
+import lottery.domain.*
 import lottery.view.InputView
 import lottery.view.InputView.printInputBonusBall
 import lottery.view.InputView.printInputLastWinnerLottery
+import lottery.view.Reception
 import lottery.view.ResultView
 import lottery.view.ResultView.printMatchNumbers
